improved delete

+++ routes/room/room.js
@@ -1,22 +1,20 @@
-const express = require('express');
-const { Room } = require('../../models/room');
-const validateRoom = require('./roomValidation');
-const mongoose = require('mongoose'); 
-const { Booking } = require('../../models/booking');
+const express = require("express");
+const { Room } = require("../../models/room");
+const validateRoom = require("./roomValidation");
+const mongoose = require("mongoose");
+const { Booking } = require("../../models/booking");
 const isClient = require("../../middleware/isClient");
 const isManager = require("../../middleware/isManager");
 
 const router = express.Router();
-
-
 
 /**
  * @swagger
  * /room:
  *   get:
  *     summary: Retrieve a list of all the rooms or available rooms by date
- *     tags: 
- *       - Rooms   
+ *     tags:
+ *       - Rooms
  *     parameters:
  *       - in: header
  *         name: x-auth-token
@@ -51,49 +49,47 @@ const router = express.Router();
  *         description: An error occurred.
  */
 
-router.get('/',isClient, async (req, res) => {
-    try {
-        const { arrivalDate, departureDate } = req.query;
+router.get("/", isClient, async (req, res) => {
+  try {
+    const { arrivalDate, departureDate } = req.query;
 
-        if(!arrivalDate && !departureDate){
-            const room = await Room.find({});
-            return res.send(room);
-        }
-        const arrival = new Date(arrivalDate);
-        const departure = new Date(departureDate);
- 
-        const bookedRooms = await Booking.find({
-            $or: [
-                {
-                    arrivalDate: { $lte: departure },
-                    departureDate: { $gte: arrival },
-                    $and: [
-                        { isApproved: { $ne: null } },
-                        { isApproved: { $ne: false } }
-                    ]
-                }
-            ]
-        }).select("roomId");
-        
-       
-        const bookedRoomIds = bookedRooms.map(booking => booking.roomId);
-
-        const availableRooms = await Room.find({
-            _id: { $nin: bookedRoomIds }
-        });
-        if (availableRooms.length === 0) return res.status(404).send('There are no available rooms in the provided dates');
-        res.send(availableRooms);
-    } catch (err) {
-      
-        res.status(500).send('An error occurred while getting the room.');
+    if (!arrivalDate && !departureDate) {
+      const room = await Room.find({});
+      return res.send(room);
     }
+    const arrival = new Date(arrivalDate);
+    const departure = new Date(departureDate);
+
+    const bookedRooms = await Booking.find({
+      $or: [
+        {
+          arrivalDate: { $lte: departure },
+          departureDate: { $gte: arrival },
+        },
+      ],
+      $and: [{ isApproved: { $ne: null } }, { isApproved: { $ne: false } }],
+    }).select("roomId");
+
+    const bookedRoomIds = bookedRooms.map((booking) => booking.roomId);
+
+    const availableRooms = await Room.find({
+      _id: { $nin: bookedRoomIds },
+    });
+    if (availableRooms.length === 0)
+      return res
+        .status(404)
+        .send("There are no available rooms in the provided dates");
+    res.send(availableRooms);
+  } catch (err) {
+    res.status(500).send("An error occurred while getting the room.");
+  }
 });
 /**
  * @swagger
  * /room/{id}:
  *   get:
  *     summary: Retrieve a room by ID
- *     tags: 
+ *     tags:
  *       - Rooms
  *     parameters:
  *       - in: header
@@ -123,27 +119,27 @@ router.get('/',isClient, async (req, res) => {
  *         description: An error occurred.
  */
 
-router.get('/:id',isManager, async (req, res) => {
-    const roomId = req.params.id;
-  
-    if (!mongoose.Types.ObjectId.isValid(roomId)) {
-        return res.status(400).send('Invalid ID format');
-    }
+router.get("/:id", isManager, async (req, res) => {
+  const roomId = req.params.id;
 
-    try {
-        const room = await Room.findById(roomId);
-        if (!room) return res.status(404).send('The room was not found');
-        res.send(room);
-    } catch (err) {
-        res.status(500).send('An error occurred while retrieving the room');
-    }
+  if (!mongoose.Types.ObjectId.isValid(roomId)) {
+    return res.status(400).send("Invalid ID format");
+  }
+
+  try {
+    const room = await Room.findById(roomId);
+    if (!room) return res.status(404).send("The room was not found");
+    res.send(room);
+  } catch (err) {
+    res.status(500).send("An error occurred while retrieving the room");
+  }
 });
 /**
  * @swagger
  * /room:
  *   post:
  *     summary: Create a new room
- *     tags: 
+ *     tags:
  *       - Rooms
  *     parameters:
  *       - in: header
@@ -167,31 +163,31 @@ router.get('/:id',isManager, async (req, res) => {
  *         description: An error occurred.
  */
 
-router.post('/',isManager,async (req, res) => {
-    try {
-        const { error } = validateRoom(req.body);
-         if (error) return res.status(400).send(error.details[0].message);
+router.post("/", isManager, async (req, res) => {
+  try {
+    const { error } = validateRoom(req.body);
+    if (error) return res.status(400).send(error.details[0].message);
 
-        const room = new Room({
-            type:req.body.type,
-            number:req.body.number,
-            description:req.body.description,
-            numberOfBeds:req.body.numberOfBeds
-        });
+    const room = new Room({
+      type: req.body.type,
+      number: req.body.number,
+      description: req.body.description,
+      numberOfBeds: req.body.numberOfBeds,
+    });
 
-        const savedRooms = await room.save();
-        
-        res.send(savedRooms);
-    } catch (err) {
-        res.status(500).send('An error occurred while saving the rooms.');
-    }
+    const savedRooms = await room.save();
+
+    res.send(savedRooms);
+  } catch (err) {
+    res.status(500).send("An error occurred while saving the rooms.");
+  }
 });
 /**
  * @swagger
  * /room/{id}:
  *   put:
  *     summary: Update an existing room
- *     tags: 
+ *     tags:
  *       - Rooms
  *     parameters:
  *       - in: header
@@ -223,36 +219,35 @@ router.post('/',isManager,async (req, res) => {
  *         description: An error occurred.
  */
 
-router.put('/:id',isManager, async (req, res) => {
+router.put("/:id", isManager, async (req, res) => {
+  const roomId = req.params.id;
+  if (!mongoose.Types.ObjectId.isValid(roomId)) {
+    return res.status(400).send("Invalid ID format");
+  }
+  try {
+    const room = await Room.findById(req.params.id);
+    if (!room) return res.status(404).send("The room was not found");
 
-    const roomId = req.params.id;
-    if (!mongoose.Types.ObjectId.isValid(roomId)) {
-        return res.status(400).send('Invalid ID format');
-    }
-    try {
-        const room = await Room.findById(req.params.id);
-        if (!room) return res.status(404).send('The room was not found');
+    const { error } = validateRoom(req.body);
+    if (error) return res.status(400).send(error.details[0].message);
 
-        const { error } = validateRoom(req.body);
-        if (error) return res.status(400).send(error.details[0].message);
-
-        room.type= req.body.type,
-        room.number=req.body.number,
-        room.description=req.body.description,
-        room.numberOfBeds=req.body.numberOfBeds
-        const updatedRoom = await room.save();
+    (room.type = req.body.type),
+      (room.number = req.body.number),
+      (room.description = req.body.description),
+      (room.numberOfBeds = req.body.numberOfBeds);
+    const updatedRoom = await room.save();
 
     res.send(updatedRoom);
-    } catch (err) {
-        res.status(500).send('An error occurred while updating the room.');
-    }
+  } catch (err) {
+    res.status(500).send("An error occurred while updating the room.");
+  }
 });
 /**
  * @swagger
  * /room/{id}:
  *   delete:
  *     summary: Delete a room by ID
- *     tags: 
+ *     tags:
  *       - Rooms
  *     parameters:
  *       - in: header
@@ -275,15 +270,28 @@ router.put('/:id',isManager, async (req, res) => {
  *       500:
  *         description: An error occurred.
  */
-router.delete('/:id',isManager, async (req, res) => {
-    try {
-       const room= await Room.findByIdAndDelete(req.params.id);
-        if (!room) return res.status(404).send('The room with this ID was not found');
-            
-        res.send(room);
-    } catch (err) {
-        res.status(500).send('An error occurred while deleting the room.');
+router.delete("/:id", isManager, async (req, res) => {
+  try {
+    const roomReferenceInBooking = await Booking.findOne({
+      roomId: req.params.id,
+    });
+
+    if (roomReferenceInBooking) {
+      return res
+        .status(409)
+        .send("Cannot delete room, because it is currently booked!");
     }
+
+    const room = await Room.findByIdAndDelete(req.params.id);
+
+    if (!room) {
+      return res.status(404).send("The room with this ID was not found");
+    }
+
+    res.send(room);
+  } catch (err) {
+    res.status(500).send("An error occurred while deleting the room.");
+  }
 });
 
- module.exports = router;
+module.exports = router;
